src/requirements.txt
discord.py>=2.5.2
python-dotenv>=0.19.0
psutil>=5.8.0
pytz>=2021.3